重构工程代码

+++ dim/src/main/java/com/singlethreadzzz/dim/service/userManage/impl/UserRoleManageServiceImpl.java
@@ -2,10 +2,13 @@ package com.singlethreadzzz.dim.service.userManage.impl;
 
 import java.util.List;
 
+import org.apache.shiro.authz.annotation.RequiresRoles;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.singlethreadzzz.dim.domain.Role;
+import com.singlethreadzzz.dim.exception.BeforeJsonException;
 import com.singlethreadzzz.dim.mapper.userManage.UseRoleManageMapper;
 import com.singlethreadzzz.dim.service.userManage.UserRoleManageService;
 import com.singlethreadzzz.dim.util.UUIDUtils;
@@ -17,38 +20,60 @@ public class UserRoleManageServiceImpl implements UserRoleManageService {
 	private UseRoleManageMapper useRoleManageMapper;
 
 	@Override
-	public Role selectUserRoleByUserId(String userId) {
+	public Role getUserRoleByUserId(String userId) throws Exception {
 		return this.useRoleManageMapper.selectUserRoleByUserId(userId);
 	}
 
 	@Override
-	public void addUserRole(Role userRole) {
+	@RequiresRoles("admin")
+	public void addUserRole(Role userRole) throws Exception {
+		Role oldUserRole = new Role();
+		oldUserRole = this.useRoleManageMapper.selectUserRoleByRoleName(userRole.getRoleName());
+		if(oldUserRole != null) {
+			throw new BeforeJsonException("角色已存在");
+		}
 		userRole.setRoleId(UUIDUtils.getUUID());
 		this.useRoleManageMapper.insertUserRole(userRole);
 	}
 
 	@Override
-	public void updateUserRole(Role userRole) {
-		this.useRoleManageMapper.updateUserByUserId(userRole);		
+	@RequiresRoles("admin")
+	public void updateUserRole(Role userRole) throws Exception {
+		Role oldUserRole = new Role();
+		oldUserRole = this.useRoleManageMapper.selectUserRoleByRoleId(userRole.getRoleId());
+		if(oldUserRole == null) {
+			throw new BeforeJsonException("角色不存在");
+		}
+		oldUserRole.setRoleCnname(userRole.getRoleCnname());
+		this.useRoleManageMapper.updateUserByUserId(oldUserRole);		
 	}
 
 	@Override
-	public void deleteUserRole(String roleId) {
-		this.useRoleManageMapper.deleteUserRoleByRoleId(roleId);		
+	@RequiresRoles("admin")
+	@Transactional
+	public void deleteUserRole(List<String> roleIdList) throws Exception {
+		for(String roleId : roleIdList) {
+			Role oldUserRole = new Role();
+			oldUserRole = this.useRoleManageMapper.selectUserRoleByRoleId(roleId);
+			if(oldUserRole == null) {
+				throw new BeforeJsonException("角色不存在");
+			}
+			this.useRoleManageMapper.deleteUserRoleByRoleId(roleId);		
+		}
 	}
 
 	@Override
-	public List<Role> getAllUserRoles() {
+	public List<Role> getAllUserRoles() throws Exception {
 		return this.useRoleManageMapper.selectAllUserRoles();
 	}
 
 	@Override
-	public Role selectUserRoleByRoleId(String roleId) {
+	public Role getUserRoleByRoleId(String roleId) throws Exception {
 		return this.useRoleManageMapper.selectUserRoleByRoleId(roleId);
 	}
 
 	@Override
-	public Role selectUserRoleByRoleName(String roleName) {
+	public Role getUserRoleByRoleName(String roleName) throws Exception {
 		return this.useRoleManageMapper.selectUserRoleByRoleName(roleName);
 	}
